Rename truffle-provider to @truffle/provider
