+++ packages/truffle-config/package.json
@@ -24,7 +24,7 @@
     "lodash": "^4.17.13",
     "original-require": "1.0.1",
     "@truffle/error": "^0.0.5",
-    "truffle-provider": "^0.1.15"
+    "@truffle/provider": "^0.1.15"
   },
   "publishConfig": {
     "access": "public"
