+++ packages/truffle-core/lib/networks.js
@@ -2,7 +2,7 @@ var fs = require("fs");
 var path = require("path");
 var OS = require("os");
 var BlockchainUtils = require("@truffle/blockchain-utils");
-var Provider = require("truffle-provider");
+var Provider = require("@truffle/provider");
 var async = require("async");
 var Web3Shim = require("truffle-interface-adapter").Web3Shim;
